CMake: Fix Sql dependency to be found when building in-tree examples
When building examples in-tree and not as external projects, the build
system sets QT_NO_CREATE_TARGETS to true.

This means that each find_package() call in an example project assumes
that the Qt target to be found already exists in scope, because the
target must have been found while looking for dependencies for building
the repo itself.

This assumption is not correct, because an example might require extra
Qt dependencies which are not needed for building the repo itself.

Because QT_NO_CREATE_TARGETS causes the QtXyzTargets.cmake file not be
included as part of a find_package() call, the package is considered
not found and it generates a very confusing message.

Work around the issue by specifying Sql to be an optional dependency
when building the repository. This will not be needed once all
repositories have migrated to building examples with ExternalProject
instead.

Fixes: QTBUG-95532
Change-Id: Ied54bc55a5846fe235a833170db733e2998afb9e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>